feat: Warning messages on wrong practices
